+++ lib/isSet.d.ts
@@ -1,0 +1,1 @@
+export declare function isSet(object: any): boolean;